+++ src/Moq/IInvocation.cs
@@ -1,6 +1,7 @@
 // Copyright (c) 2007, Clarius Consulting, Manas Technology Solutions, InSTEDD, and Contributors.
 // All rights reserved. Licensed under the BSD 3-Clause License; see License.txt.
 
+using System;
 using System.Collections.Generic;
 using System.Reflection;
 
@@ -32,8 +33,13 @@ namespace Moq
 		bool IsVerified { get; }
 
 		/// <summary>
-		/// Gets the return value of the invocation.
+		/// The value being returned for a non-void method if no exception was thrown.
 		/// </summary>
 		object ReturnValue { get; }
+
+		/// <summary>
+		/// Optional exception if the method invocation results in an exception being thrown.
+		/// </summary>
+		Exception Exception { get; }
 	}
 }